* change: [API] rgb<uint8_t> use 655 bit partition instead 556 * fix: [API] broken gray<uint*_t>, on component, bound and container

+++ src/color/gray/trait/component/uint64.hpp
@@ -16,7 +16,7 @@ namespace color
 
     template< >
      struct component< ::color::category::gray_uint64 >
-      : public ::color::_internal::utility::component::Unsigned< std::uint16_t, unsigned>
+      : public ::color::_internal::utility::component::Unsigned< std::uint64_t, unsigned>
       {
       };
